requirements.txt
# iflow2web dependencies
fastapi==0.115.0
uvicorn[standard]==0.32.0
websockets==13.1
iflow-cli-sdk==0.2.1
jinja2==3.1.4
python-dotenv==1.0.1

# Test dependencies
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2